// frontend/src/api/generated/models/api-key-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * chainlink-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiKeyJson
 */
export interface ApiKeyJson {
    /**
     * 
     * @type {string}
     * @memberof ApiKeyJson
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyJson
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKeyJson
     */
    prefix: string;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyJson
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyJson
     */
    expiresAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyJson
     */
    lastUsedAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ApiKeyJson
     */
    revokedAt?: Date;
}

/**
 * Check if a given object implements the ApiKeyJson interface.
 */
export function instanceOfApiKeyJson(value: object): value is ApiKeyJson {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('prefix' in value) || value['prefix'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function ApiKeyJsonFromJSON(json: any): ApiKeyJson {
    return ApiKeyJsonFromJSONTyped(json, false);
}

export function ApiKeyJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiKeyJson {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'prefix': json['prefix'],
        'createdAt': (new Date(json['createdAt'])),
        'expiresAt': json['expiresAt'] == null ? undefined : (new Date(json['expiresAt'])),
        'lastUsedAt': json['lastUsedAt'] == null ? undefined : (new Date(json['lastUsedAt'])),
        'revokedAt': json['revokedAt'] == null ? undefined : (new Date(json['revokedAt'])),
    };
}

export function ApiKeyJsonToJSON(json: any): ApiKeyJson {
    return ApiKeyJsonToJSONTyped(json, false);
}

export function ApiKeyJsonToJSONTyped(value?: ApiKeyJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'prefix': value['prefix'],
        'createdAt': value['createdAt'].toISOString(),
        'expiresAt': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'lastUsedAt': value['lastUsedAt'] == null ? value['lastUsedAt'] : value['lastUsedAt'].toISOString(),
        'revokedAt': value['revokedAt'] == null ? value['revokedAt'] : value['revokedAt'].toISOString(),
    };
}